feat(api): wire MCP search/execute through clawql-api adapters
Route search and execute MCP handlers via Effect SearchService/ExecuteService
layers, extracting execute core logic for strangler migration into clawql-api.
